Add function for check file with todos

+++ card/cards.post.js
@@ -11,27 +11,35 @@ const filePath = 'cards.json'
 const router = Router.post('/', body('name').isString(), function(req,res){
     if(!req.body.name) return res.status(400).send("Dont name task")
     
-        const content  = fs.readFileSync(filePath, 'utf8')     
-        const cards = JSON.parse(content)
-
         const errors = validationResult(req);
         if (!errors.isEmpty()) {
           return res.status(400).json({ errors: errors.array() });
         }
-        
+
         const newCards = {uuid: Math.floor(Math.random()*1000000000000000), 
             date: new Date(), 
             done:false, 
             name: req.body.name}
 
-        cards.push(newCards)
+        if(fs.existsSync(filePath)){
+            const content  = fs.readFileSync(filePath, 'utf8')     
+            const cards = JSON.parse(content)
+
+            cards.push(newCards)
 
             fs.writeFileSync(filePath, JSON.stringify(cards), err => {
                 if(err){
-                    res.send(err)
+                     res.send(err)
                 }
             }) 
-        res.send(newCards)        
+            res.send(newCards)  
+        }
+        else {
+            fs.appendFile(filePath, JSON.stringify([newCards]), function (err) {
+                if (err) throw err
+                res.send(newCards) 
+            })
+        }
 })
 
 module.exports = router
